Refactor assets folder
